# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api-keys": {
      "post": {
        "operationId": "ApiKeysController_createApiKey",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "required": false,
            "in": "header",
            "description": "Tenant ID (required in multi-tenant mode)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateApiKeyResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing required parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Create a new API key",
        "tags": [
          "api-keys"
        ]
      },
      "get": {
        "operationId": "ApiKeysController_getApiKeys",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "required": false,
            "in": "header",
            "description": "Tenant ID (required in multi-tenant mode)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API keys retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiKey"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - tenantId required in multi-tenant mode"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get all API keys for a tenant",
        "tags": [
          "api-keys"
        ]
      }
    },
    "/api-keys/{id}": {
      "get": {
        "operationId": "ApiKeysController_getApiKey",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "API key ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-ID",
            "required": false,
            "in": "header",
            "description": "Tenant ID (required in multi-tenant mode)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKey"
                }
              }
            }
          },
          "404": {
            "description": "API key not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get a specific API key by ID",
        "tags": [
          "api-keys"
        ]
      },
      "put": {
        "operationId": "ApiKeysController_updateApiKey",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "API key ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-ID",
            "required": false,
            "in": "header",
            "description": "Tenant ID (required in multi-tenant mode)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApiKeyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "API key updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKey"
                }
              }
            }
          },
          "404": {
            "description": "API key not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Update an API key",
        "tags": [
          "api-keys"
        ]
      },
      "delete": {
        "operationId": "ApiKeysController_deleteApiKey",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "API key ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-ID",
            "required": false,
            "in": "header",
            "description": "Tenant ID (required in multi-tenant mode)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deleted": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "API key not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Delete an API key",
        "tags": [
          "api-keys"
        ]
      }
    },
...
    "/extraction/extract": {
      "post": {
        "operationId": "ExtractionController_extractPricingTable",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "required": false,
            "in": "header",
            "description": "Tenant ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PricingTableExtractionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pricing table extracted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractedPricingTable"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractedPricingTable"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid image data"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Extract pricing table from image",
        "tags": [
          "extraction"
        ]
      }
    },
...
    "/auth/users": {
      "post": {
        "operationId": "AuthController_ensureUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User created or retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          }
        },
        "summary": "Create or get user by Firebase ID",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Pricing Agent Builder API",
    "description": "API documentation for the Pricing Agent Builder service",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateApiKeyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "ApiKey": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "isActive": {
            "type": "boolean",
            "default": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "lastUsedAt": {
            "format": "date-time",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "name",
          "key",
          "tenantId",
          "isActive",
          "createdAt"
        ]
      },
      "CreateApiKeyResponseDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "apiKey": {
            "$ref": "#/components/schemas/ApiKey"
          }
        },
        "required": [
          "key",
          "apiKey"
        ]
      },
      "UpdateApiKeyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "expiresAt": {
            "type": "string"
          }
        }
      },
...
      "PricingTableExtractionRequest": {
        "type": "object",
        "properties": {
          "imageData": {
            "type": "string"
          },
          "imageMimeType": {
            "type": "string"
          },
          "additionalContext": {
            "type": "string"
          }
        },
        "required": [
          "imageData",
          "imageMimeType"
        ]
      },
      "ExtractedPricingTable": {
        "type": "object",
        "properties": {
          "markdown": {
            "type": "string"
          },
          "rawText": {
            "type": "string"
          },
          "confidence": {
            "type": "number"
          }
        },
        "required": [
          "markdown",
          "rawText",
          "confidence"
        ]
      },
...
      "UserDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "format": "uuid"
          },
          "firebaseId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "selectedTenantId": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean",
            "default": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "firebaseId",
          "email",
          "isActive",
          "createdAt"
        ]
      },
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "firebaseId": {
            "type": "string",
            "description": "Firebase user ID"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email"
          },
          "role": {
            "type": "string",
            "description": "User role",
            "default": "admin"
          }
        },
        "required": [
          "firebaseId",
          "email"
        ]
      }
    }
  }
}